+++ CHANGELOG.md
@@ -27,10 +27,19 @@ This is an example of a brief overview of the _Major_ or _Minor_ version changes
 
 ---
 
-## [9.1] `DateInput` popover behavior
+## [9.3] `DateInput` popover behavior
 
 - `DateInput` now opens the calendar popover on input focus
 - The calendar popover in `DateInput` will now close itself on date selection
+
+## [9.2] Avatar enhancements
+
+- Added `l` size to Avatar
+- Added `PlaceholderIcon` prop
+
+## [9.1] TextInput enhancements
+
+- Added `labelWidth` prop to control width of label (helpful when used in combination with `labelPosition`)
 
 ## [9.0] MAJOR